% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd.meddra}
\alias{dd.meddra}
\title{Medical Dictionary for Regulatory Activities (MedDRA)}
\format{
An object of class \code{data.frame} with 12482 rows and 2 columns.
}
\usage{
dd.meddra
}
\description{
This example data uses the HTML level terms of MedDRA
to define AE groups. AE groups may overlap slightly.
}
\details{
The data contains 2 columns
\itemize{
  \item AE_NAME Adverse event name
  \item GROUP_NAME AE group name
}
}
\keyword{datasets}
