% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merge_list}
\alias{merge_list}
\title{An example of merge list}
\format{
An object of class \code{list} of length 10.
}
\usage{
merge_list
}
\description{
The function `zinb_analysis_tool` requires users to define
the vaccines of interest and which vaccines are merged in to
one vaccine type. Here is an example of how to specify the
parameter `merge_list` in that function.
}
\details{
The `merge_list` should be a list of length i. An element of
`merge_list` should be a list of length 2, with the first element
to be an vector string containing one or several vaccine names,
and the second element to be a single string of vaccine type.

For example, an element of `merge_list` could be `list(c('FLUN3', 'FLUN4'),'FLUN')`,
meaning to merge FLUN3 and FLUN4 into one vaccine type FLUN.
}
\keyword{datasets}
