% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.zGPS}
\alias{plot.zGPS}
\title{Generate the heat map or the scatter plot of lambdas}
\usage{
\method{plot}{zGPS}(x, vaccine_name = NULL, AE_grp_name = NULL, interactive_plot = FALSE, ...)
}
\arguments{
\item{x}{An object of class zGPS}

\item{vaccine_name}{The vaccine type to be visualized}

\item{AE_grp_name}{The AE group name to be visualized}

\item{interactive_plot}{Interactive plots, default to false}

\item{...}{more arguments}
}
\value{
The plot of the model
}
\description{
Generate the heat map or the scatter plot of lambdas
}
