% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generate the heatmap}
\usage{
plot_heatmap(big_data)
}
\arguments{
\item{big_data}{The big data.frame from the `zinb_analysis_tool` function}
}
\value{
A heatmap produced by `ggplot2::ggplot()`
}
\description{
This function is a visualization tool of group level RRs
obtained from the `zinb_analysis_tool`. It will generate a
heatmap with each cell representing the
RR of an AE group with a vaccine.
}
