% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lambdas.R
\name{plot_lambdas}
\alias{plot_lambdas}
\title{Generate the scatter plot of lambdas}
\usage{
plot_lambdas(vaccine_name, AE_grp_name, big_data)
}
\arguments{
\item{vaccine_name}{The vaccine type to be visualized}

\item{AE_grp_name}{The AE group name to be visualized}

\item{big_data}{The big data.frame from the `zinb_analysis_tool` function}
}
\value{
A scatter plot produced by `ggplot2::ggplot()`
}
\description{
This function is to visualize the AE level RRs in an AE group.
}
