% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_heatmap.R
\name{plotly_heatmap}
\alias{plotly_heatmap}
\title{Generate the heatmap (Plotly)}
\usage{
plotly_heatmap(big_data)
}
\arguments{
\item{big_data}{The big data.frame from the `zinb_analysis_tool` function}
}
\value{
A heatmap produced by `plotly::plotly()`
}
\description{
This function is an advanced visualization tool of group level RRs
obtained from the `zinb_analysis_tool`. It will generate an interactive
heatmap with each cell representing the
RR of an AE group with a vaccine.
}
