% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaers_data}
\alias{vaers_data}
\title{The VAERS data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1017599 rows and 3 columns.
}
\usage{
vaers_data
}
\description{
VAERS stands for the Vaccine Adverse Event Reporting System.
This system was created by the Food and Drug Administration (FDA) and
Centers for Disease Control and Prevention (CDC)
to receive reports about adverse events that may be associated with vaccines.
}
\details{
This example data contains VAERS reports from 2005-01-01 to 2018-12-31 with
the patient age ranging from 2 to 49.

This are three columns in this data:
\itemize{
  \item ID The unique identifier of each report in VAERS data
  \item VAX_TYPE Vaccine type(s) in the report
  \item AE_NAME Adverse event(s) in the report
}
}
\keyword{datasets}
