% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zGPS_bootstrap.R
\name{zGPS_bootstrap}
\alias{zGPS_bootstrap}
\title{Perform bootstrap to evaluate accuracy of zGPS analysis.}
\usage{
zGPS_bootstrap(zGPS_result, n_perm = 20, n_cores = round(detectCores()/2))
}
\arguments{
\item{n_perm}{a positive integer that decide the number of bootstrap samples to determine
the p value for observed RRs.}

\item{n_copies}{a positive integer indicating the number of permutations can
be done in parallel. This number is default to half of the computer's number
of cores.}
}
\value{
An object of S3 class zGPS.
It contain the bootstrap results and the calculated q values.
}
\description{
Perform bootstrap to evaluate accuracy of zGPS analysis.
}
